+++ releasenotes.md
@@ -2,3 +2,4 @@ Whats New:
 ----------------------
 Include compositions in Where Used page
 Add solvents and carriers as a separate category
+We highlight matching patterns when searching in formula